[Chore] #159 - 메모 텍스트뷰 컴포넌트화 #1
