+++ Peekabook/Peekabook/Presentation/UIComponents/MemoView.swift
@@ -1,0 +1,105 @@
+//
+//  MemoView.swift
+//  Peekabook
+//
+//  Created by 고두영 on 2023/02/23.
+//
+
+import UIKit
+
+class MemoView: UIView {
+    
+    // MARK: - UI Components
+    
+    let memoBoxView = UIView()
+    private let memoHeaderView = UIView()
+    
+    private let memoLabel = UILabel().then {
+        $0.text = I18N.BookDetail.memo
+        $0.font = .h1
+        $0.textColor = .peekaWhite
+    }
+    
+    let memoTextView = UITextView().then {
+        $0.font = .h2
+        $0.textColor = .peekaGray1
+        $0.text = I18N.BookDetail.memoHint
+        $0.backgroundColor = .clear
+        $0.autocorrectionType = .no
+        $0.textContainerInset = .init(top: 0, left: -5, bottom: 0, right: 0)
+        $0.returnKeyType = .done
+    }
+    
+    let memoMaxLabel = UILabel().then {
+        $0.text = "0/50"
+        $0.font = .h2
+        $0.textColor = .peekaGray2
+    }
+    
+    // MARK: - Initialization
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setUI()
+        setLayout()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
+
+// MARK: - Methods
+
+extension MemoView {
+    
+    private func setUI() {
+        backgroundColor = .clear
+        memoBoxView.backgroundColor = .peekaWhite_60
+        memoBoxView.layer.borderWidth = 2
+        memoBoxView.layer.borderColor = UIColor.peekaRed.cgColor
+        memoHeaderView.backgroundColor = .peekaRed
+    }
+    
+    private func setLayout() {
+        
+        addSubviews(memoBoxView, memoMaxLabel)
+        
+        [memoHeaderView, memoTextView].forEach {
+            memoBoxView.addSubview($0)
+        }
+        
+        [memoLabel].forEach {
+            memoHeaderView.addSubview($0)
+        }
+        
+        memoBoxView.snp.makeConstraints { make in
+            make.top.equalToSuperview()
+            make.centerX.equalToSuperview()
+            make.width.equalTo(335)
+            make.height.equalTo(101)
+        }
+        
+        memoHeaderView.snp.makeConstraints { make in
+            make.top.equalToSuperview()
+            make.centerX.equalToSuperview()
+            make.width.equalToSuperview()
+            make.height.equalTo(36)
+        }
+        
+        memoLabel.snp.makeConstraints { make in
+            make.centerY.equalToSuperview()
+            make.leading.equalToSuperview().offset(14)
+        }
+        
+        memoTextView.snp.makeConstraints { make in
+            make.top.equalTo(memoHeaderView.snp.bottom).offset(10)
+            make.leading.trailing.bottom.equalTo(memoBoxView).inset(14)
+        }
+        
+        memoMaxLabel.snp.makeConstraints { make in
+            make.top.equalTo(memoBoxView.snp.bottom).offset(8)
+            make.trailing.equalTo(memoBoxView.snp.trailing)
+        }
+    }
+}
